# server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "Manager API",
        "contact": {}
    },
    "host": "localhost:3000",
    "basePath": "/v1",
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "scheme": "bearer"
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/role-master/active-deactive": {
            "get": {
                "summary": "active-deactive",
                "tags": [
                    "Role Master"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "operationId": "active-deactive",
                "deprecated": false,
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "_id",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "description": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                }
            }
        },
...
        "/location-master/activeList": {
            "get": {
                "summary": "activeList",
                "tags": [
                    "Location Master"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "operationId": "activeList",
                "deprecated": false,
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                }
            }
        },
...
    },
...
}